+++ NEWS.md
@@ -1,3 +1,8 @@
+# Development version
+* Improved the manual.
+* Added `sim_fun_test2` function for testing.
+* Added `...` for `batch_simulation`.
+
 # simlandr 0.1.1
 * Cleaned the package to make it CRAN-compatible.